Added matlab.mixin.Copyable back in, but marked lines that must be replaced (3 lines) in dj.GeneralRelvar to make it R2009 compatible.
If the server disconnects while in an ongoing transaction, dj.Connection now prints an explanation and hands the control to the user (keyboard). If the user types dbcont at the prompt, the rest of the transaction will still be executed. Normally, users should dbquit and restart the processing. This change is necessary for very long transactions when only one item is computed but it takes so long that the server may disconnect during the makeTuples call.



git-svn-id: http://datajoint.googlecode.com/svn/trunk@499 2f38a6b3-5502-11bb-9305-b9e7ad8f4d38